pyproject.toml
[project]
name = "maptr-study"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "einops>=0.8.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "nuscenes-devkit>=1.1.9",
    "pandas>=2.3.3",
    "scipy>=1.16.3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ruff>=0.5.0",
    "opencv-python>=4.11.0.86",
    "tensorboard>=2.20.0",
]
